+++ packages/cli/src/bin.ts
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 import { main } from "./cli.js";
 
-main();
+void main();